Migration attribute supports longs for version number.

+++ Raven.Migrations/MigrationAttribute.cs
@@ -7,13 +7,13 @@ namespace Raven.Migrations
     [AttributeUsage(AttributeTargets.Class, AllowMultiple = false, Inherited = false)]
     public class MigrationAttribute : Attribute
     {
-        public MigrationAttribute(int version)
+        public MigrationAttribute(long version)
             : this(version, null)
         {
             
         }
 
-        public MigrationAttribute(int version, params string[] profiles)
+        public MigrationAttribute(long version, params string[] profiles)
         {
             Version = version;
             Profiles = profiles ?? Enumerable.Empty<string>();
